+++ server/src-lib/Hasura/App.hs
@@ -525,6 +525,7 @@ runHGEServer setupHook env ServeOptions{..} ServeCtx{..} initTime postPollHook s
     fetchI        = milliseconds $ fromMaybe (Milliseconds defaultFetchInterval) soEventsFetchInterval
     logEnvHeaders = soLogHeadersFromEnv
     allPgSources  = mapMaybe (unsafeSourceConfiguration @('Postgres 'Vanilla)) $ HM.elems $ scSources $ lastBuiltSchemaCache _scSchemaCache
+    eventResponseLogBehaviour = if soInDevelopmentMode then LogEntireResponse else LogSanitisedResponse
 
   -- TODO: is this correct?
   -- event triggers should be tied to the life cycle of a source
@@ -550,6 +551,7 @@ runHGEServer setupHook env ServeOptions{..} ServeCtx{..} initTime postPollHook s
                         lockedEventsCtx
                         serverMetrics
                         soEnableMaintenanceMode
+                        eventResponseLogBehaviour
 
   -- start a backgroud thread to handle async actions
   case soAsyncActionsFetchInterval of
@@ -574,6 +576,7 @@ runHGEServer setupHook env ServeOptions{..} ServeCtx{..} initTime postPollHook s
   _scheduledEventsThread <- C.forkManagedT "processScheduledTriggers" logger $
     processScheduledTriggers env logger logEnvHeaders _scHttpManager
                              (getSCFromRef cacheRef) lockedEventsCtx
+                             eventResponseLogBehaviour
 
   -- start a background thread to check for updates
   _updateThread <- C.forkManagedT "checkForUpdates" logger $ liftIO $