+++ server/src-lib/Hasura/Eventing/EventTrigger.hs
@@ -39,6 +39,7 @@ module Hasura.Eventing.EventTrigger
   , Event(..)
   , unlockEvents
   , EventEngineCtx(..)
+  , ResponseLogBehavior(..)
   ) where
 
 import           Hasura.Prelude
@@ -236,8 +237,9 @@ processEventQueue
   -> LockedEventsCtx
   -> ServerMetrics
   -> MaintenanceMode
+  -> ResponseLogBehavior
   -> m void
-processEventQueue logger logenv httpMgr getSchemaCache EventEngineCtx{..} LockedEventsCtx{leEvents} serverMetrics maintenanceMode = do
+processEventQueue logger logenv httpMgr getSchemaCache EventEngineCtx{..} LockedEventsCtx{leEvents} serverMetrics maintenanceMode responseLogBehavior = do
   events0 <- popEventsBatch
   -- Track number of events fetched in EKG
   _ <- liftIO $ EKG.Distribution.add (smNumEventsFetched serverMetrics) (fromIntegral $ length events0)
@@ -409,7 +411,7 @@ processEventQueue logger logenv httpMgr getSchemaCache EventEngineCtx{..} Locked
                       (liftIO $ EKG.Gauge.inc $ smNumEventHTTPWorkers serverMetrics)
                       (liftIO $ EKG.Gauge.dec $ smNumEventHTTPWorkers serverMetrics)
                       (runExceptT $ tryWebhook headers responseTimeout payload webhook)
-              logHTTPForET res extraLogCtx requestDetails
+              logHTTPForET res extraLogCtx requestDetails responseLogBehavior
               let decodedHeaders = map (decodeHeader logenv headerInfos) headers
               either
                 (processError sourceConfig e retryConf decodedHeaders ep maintenanceModeVersion)